+++ CHANGELOG.md
@@ -1,3 +1,232 @@
+## [1.0.0-next.36](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.35...v1.0.0-next.36) (2025-08-25)
+
+### Features
+
+* add default return key type support to NitroTextInput ([#46](https://github.com/Ucekay/react-native-nitro-text-input/issues/46)) ([669ab71](https://github.com/Ucekay/react-native-nitro-text-input/commit/669ab7110093f0e36138ad6fa4bd343c22128e7d))
+
+## [1.0.0-next.35](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.34...v1.0.0-next.35) (2025-08-24)
+
+### Features
+
+* add focus, blur, clear, and isFocused methods to NitroTextInput ([#45](https://github.com/Ucekay/react-native-nitro-text-input/issues/45)) ([b397201](https://github.com/Ucekay/react-native-nitro-text-input/commit/b397201c6ea8d86cae12013f952589203f8ff8fc))
+
+## [1.0.0-next.34](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.33...v1.0.0-next.34) (2025-08-23)
+
+### Features
+
+* add password rules support ([#44](https://github.com/Ucekay/react-native-nitro-text-input/issues/44)) ([1f8073b](https://github.com/Ucekay/react-native-nitro-text-input/commit/1f8073beb5ca0578c3c071e3c4a0d129c6138fb2))
+
+## [1.0.0-next.33](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.32...v1.0.0-next.33) (2025-08-23)
+
+### Features
+
+* add textAlign prop ([#43](https://github.com/Ucekay/react-native-nitro-text-input/issues/43)) ([ef94dc5](https://github.com/Ucekay/react-native-nitro-text-input/commit/ef94dc5f5e8ba44413909cd61ed5bb22ecb58a17))
+
+## [1.0.0-next.32](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.31...v1.0.0-next.32) (2025-08-23)
+
+### Features
+
+* add submitBehavior prop ([#42](https://github.com/Ucekay/react-native-nitro-text-input/issues/42)) ([2a52ee6](https://github.com/Ucekay/react-native-nitro-text-input/commit/2a52ee673660d24a75009d6deb05532c57da4e2b))
+
+## [1.0.0-next.31](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.30...v1.0.0-next.31) (2025-08-23)
+
+### Features
+
+* add spellCheck support ([#41](https://github.com/Ucekay/react-native-nitro-text-input/issues/41)) ([48cfc63](https://github.com/Ucekay/react-native-nitro-text-input/commit/48cfc63e1bef7939c7d3afdee92f4026a02b9538))
+
+## [1.0.0-next.30](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.29...v1.0.0-next.30) (2025-08-23)
+
+### Features
+
+* add smartInsertDelete feature for iOS text input ([#40](https://github.com/Ucekay/react-native-nitro-text-input/issues/40)) ([1930913](https://github.com/Ucekay/react-native-nitro-text-input/commit/193091357e79cb9417c04f49e30e6d74bc0b24fa))
+
+## [1.0.0-next.29](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.28...v1.0.0-next.29) (2025-08-23)
+
+### Features
+
+* add showSoftInputOnFocus property for iOS text input ([#39](https://github.com/Ucekay/react-native-nitro-text-input/issues/39)) ([7a1c6f3](https://github.com/Ucekay/react-native-nitro-text-input/commit/7a1c6f3610d865b750451142be0bca3fcf07df15))
+
+## [1.0.0-next.28](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.27...v1.0.0-next.28) (2025-08-23)
+
+### Features
+
+* add selectTextOnFocus property ([#38](https://github.com/Ucekay/react-native-nitro-text-input/issues/38)) ([7bc3c91](https://github.com/Ucekay/react-native-nitro-text-input/commit/7bc3c918589dcb23ae3b45acaf0bb9f93dcd5a4a))
+
+## [1.0.0-next.27](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.26...v1.0.0-next.27) (2025-08-23)
+
+### Features
+
+* add selectionColor property ([#37](https://github.com/Ucekay/react-native-nitro-text-input/issues/37)) ([ec79469](https://github.com/Ucekay/react-native-nitro-text-input/commit/ec79469f92dcfa8caca8413e7fb0aa44a304a61f))
+
+## [1.0.0-next.26](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.25...v1.0.0-next.26) (2025-08-23)
+
+### Features
+
+* add selection prop ([#36](https://github.com/Ucekay/react-native-nitro-text-input/issues/36)) ([6abdcab](https://github.com/Ucekay/react-native-nitro-text-input/commit/6abdcabc5c494dd1ec558ef93ab5604e3c7c06eb))
+
+## [1.0.0-next.25](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.24...v1.0.0-next.25) (2025-08-23)
+
+### Features
+
+* add secureTextEntry support for iOS text input ([#35](https://github.com/Ucekay/react-native-nitro-text-input/issues/35)) ([b306bd0](https://github.com/Ucekay/react-native-nitro-text-input/commit/b306bd02954577028339151f84155dc8932961ec))
+
+## [1.0.0-next.24](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.23...v1.0.0-next.24) (2025-08-23)
+
+### Bug Fixes
+
+* add react-native-nitro-modules patch and refactor NitroTextInput types ([#34](https://github.com/Ucekay/react-native-nitro-text-input/issues/34)) ([27f54a7](https://github.com/Ucekay/react-native-nitro-text-input/commit/27f54a73f41b3034fae96c7d147193a1d83ee386))
+
+## [1.0.0-next.23](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.22...v1.0.0-next.23) (2025-08-23)
+
+### Features
+
+* add returnKeyType support for NitroTextInput ([#33](https://github.com/Ucekay/react-native-nitro-text-input/issues/33)) ([c488102](https://github.com/Ucekay/react-native-nitro-text-input/commit/c4881025110cc6a7a1e3f57b4ee36ac3e8668f0f))
+
+## [1.0.0-next.22](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.21...v1.0.0-next.22) (2025-08-23)
+
+### Features
+
+* add keyboardAppearance prop to TextInput ([#31](https://github.com/Ucekay/react-native-nitro-text-input/issues/31)) ([d1ace7b](https://github.com/Ucekay/react-native-nitro-text-input/commit/d1ace7b98a171f29737ac7fabd89eb80f814dc03))
+
+## [1.0.0-next.21](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.20...v1.0.0-next.21) (2025-08-23)
+
+### Features
+
+* add placeholderTextColor support for iOS TextInput ([#30](https://github.com/Ucekay/react-native-nitro-text-input/issues/30)) ([6bcfed7](https://github.com/Ucekay/react-native-nitro-text-input/commit/6bcfed79eca528ddf8f35cbef2cddab8d55c4d0d))
+
+## [1.0.0-next.20](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.19...v1.0.0-next.20) (2025-08-22)
+
+### Features
+
+* **text-input:** add onSubmitEditing support (iOS) ([61c5df7](https://github.com/Ucekay/react-native-nitro-text-input/commit/61c5df76bccf6456361948c61dcef0cf81682d1e))
+
+## [1.0.0-next.19](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.18...v1.0.0-next.19) (2025-08-22)
+
+### Features
+
+* add onSelectionChange support for NitroTextInput ([#28](https://github.com/Ucekay/react-native-nitro-text-input/issues/28)) ([80eebd6](https://github.com/Ucekay/react-native-nitro-text-input/commit/80eebd67e0b6167cd1ddc78983b5f8d738dedad9))
+
+## [1.0.0-next.18](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.17...v1.0.0-next.18) (2025-08-22)
+
+### Features
+
+* add onKeyPress callback to NitroTextInput ([#25](https://github.com/Ucekay/react-native-nitro-text-input/issues/25)) ([6714a2c](https://github.com/Ucekay/react-native-nitro-text-input/commit/6714a2c7dda8cb2397e7f95858ce4fc39971ed90))
+
+## [1.0.0-next.17](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.16...v1.0.0-next.17) (2025-08-22)
+
+### Features
+
+* add onFocus callback to NitroTextInput component ([#24](https://github.com/Ucekay/react-native-nitro-text-input/issues/24)) ([6048f11](https://github.com/Ucekay/react-native-nitro-text-input/commit/6048f11cc70fecd6588a9cbaf95b4ff529868c5c))
+
+## [1.0.0-next.16](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.15...v1.0.0-next.16) (2025-08-21)
+
+### Features
+
+* add implementation guidelines and TODO tracking for features ([5a838df](https://github.com/Ucekay/react-native-nitro-text-input/commit/5a838df902186cc06e683b0c2d8f520acc4404ad))
+
+## [1.0.0-next.15](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.14...v1.0.0-next.15) (2025-08-21)
+
+### Features
+
+* add onTouchEnded (onPressOut) aupport ([#23](https://github.com/Ucekay/react-native-nitro-text-input/issues/23)) ([06401fb](https://github.com/Ucekay/react-native-nitro-text-input/commit/06401fb98d353e9f407085a9ab6eb708cdc67866))
+
+## [1.0.0-next.14](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.13...v1.0.0-next.14) (2025-08-20)
+
+### Features
+
+* add onTouchBegan (onPressIn) support to NitroTextInput components ([#21](https://github.com/Ucekay/react-native-nitro-text-input/issues/21)) ([007dab6](https://github.com/Ucekay/react-native-nitro-text-input/commit/007dab642c21030dee9a0a38839f7a6f91c1d66a))
+
+## [1.0.0-next.13](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.12...v1.0.0-next.13) (2025-08-20)
+
+### Features
+
+* add onEditingEnded callback to NitroTextInput component ([#20](https://github.com/Ucekay/react-native-nitro-text-input/issues/20)) ([384d208](https://github.com/Ucekay/react-native-nitro-text-input/commit/384d2085210f8cd39446ccd1203048a112bd09db))
+
+## [1.0.0-next.12](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.11...v1.0.0-next.12) (2025-08-20)
+
+### Features
+
+* add onTextChanged callback to NitroTextInput components ([#19](https://github.com/Ucekay/react-native-nitro-text-input/issues/19)) ([c101d30](https://github.com/Ucekay/react-native-nitro-text-input/commit/c101d30d964a5fd94013895aa1710eb507247544))
+
+## [1.0.0-next.11](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.10...v1.0.0-next.11) (2025-08-20)
+
+### Features
+
+* add onBlurred and onDidEndEditing callbacks to text input components ([#18](https://github.com/Ucekay/react-native-nitro-text-input/issues/18)) ([253b7b7](https://github.com/Ucekay/react-native-nitro-text-input/commit/253b7b7556f9c974f3f95f3b45878b12a3804184))
+
+## [1.0.0-next.10](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.9...v1.0.0-next.10) (2025-08-20)
+
+### Features
+
+* add maxLength support to NitroTextInput components ([#17](https://github.com/Ucekay/react-native-nitro-text-input/issues/17)) ([3c46588](https://github.com/Ucekay/react-native-nitro-text-input/commit/3c465885341e0e818b6cd71333c3b01c0715fde4))
+
+## [1.0.0-next.9](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.8...v1.0.0-next.9) (2025-08-20)
+
+### Features
+
+* add maxFontSizeMultiplier support to NitroTextInput components ([fe7323a](https://github.com/Ucekay/react-native-nitro-text-input/commit/fe7323a3631332a2bf37d32dde7b32de82e80c1a))
+* **text-input:** add maxFontSizeMultiplier support for font scaling ([d5455af](https://github.com/Ucekay/react-native-nitro-text-input/commit/d5455af27c616803730bfb033be8cceee2cab529))
+
+## [1.0.0-next.8](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.7...v1.0.0-next.8) (2025-08-20)
+
+### Features
+
+* add keyboardType and inputMode support to text inputs ([#16](https://github.com/Ucekay/react-native-nitro-text-input/issues/16)) ([554a181](https://github.com/Ucekay/react-native-nitro-text-input/commit/554a1818c3167d6a364f060299148b2f18145586))
+
+## [1.0.0-next.7](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.6...v1.0.0-next.7) (2025-08-19)
+
+### Features
+
+* add EnterKeyHint support to NitroTextInput components ([#15](https://github.com/Ucekay/react-native-nitro-text-input/issues/15)) ([bbb6410](https://github.com/Ucekay/react-native-nitro-text-input/commit/bbb6410a3a256c4f65a43675b52e428a9215c7f4))
+
+## [1.0.0-next.6](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.5...v1.0.0-next.6) (2025-08-19)
+
+### Features
+
+* add new properties and enhancements to NitroTextInput components ([#14](https://github.com/Ucekay/react-native-nitro-text-input/issues/14)) ([a39e3c0](https://github.com/Ucekay/react-native-nitro-text-input/commit/a39e3c0beb7d21c74320416c73465ef9ebea494e))
+
+## [1.0.0-next.5](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.4...v1.0.0-next.5) (2025-08-19)
+
+### Features
+
+* add editable property to NitroTextInput and HybridTextInputView ([#12](https://github.com/Ucekay/react-native-nitro-text-input/issues/12)) ([b9d127c](https://github.com/Ucekay/react-native-nitro-text-input/commit/b9d127c969c283bde07aee0cde566f75dda733cd))
+
+## [1.0.0-next.4](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.3...v1.0.0-next.4) (2025-08-19)
+
+### Features
+
+* add new properties to NitroTextInput components ([#11](https://github.com/Ucekay/react-native-nitro-text-input/issues/11)) ([a30c67f](https://github.com/Ucekay/react-native-nitro-text-input/commit/a30c67fab309e244549b24f6c28f3bd51ac93510))
+
+## [1.0.0-next.3](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.2...v1.0.0-next.3) (2025-08-19)
+
+### Features
+
+* add defaultValue prop to NitroTextInput component ([#10](https://github.com/Ucekay/react-native-nitro-text-input/issues/10)) ([6d4d44d](https://github.com/Ucekay/react-native-nitro-text-input/commit/6d4d44ddc550130d4e3628682106d7e26a89c181))
+
+## [1.0.0-next.2](https://github.com/Ucekay/react-native-nitro-text-input/compare/v1.0.0-next.1...v1.0.0-next.2) (2025-08-19)
+
+### Features
+
+* Added a new prop `clearTextOnFocus` ([#9](https://github.com/Ucekay/react-native-nitro-text-input/issues/9)) ([3a4f1e1](https://github.com/Ucekay/react-native-nitro-text-input/commit/3a4f1e16b3aff94e6f8c4b93828eedb7ca21c92a))
+
+## 1.0.0-next.1 (2025-08-19)
+
+### Features
+
+* Add support for auto capitalization and font scaling in NitroTextInput ([15b97c8](https://github.com/Ucekay/react-native-nitro-text-input/commit/15b97c8675f334e6646a171a303e5f1ab98a243c))
+* **app:** update app structure and dependencies ([8721326](https://github.com/Ucekay/react-native-nitro-text-input/commit/8721326e79502ec48fffd6b6b88ef33a68d8f634))
+* Initialize Expo project with TypeScript ([f1c0344](https://github.com/Ucekay/react-native-nitro-text-input/commit/f1c034418cd87ac96471afb781219d80fd70b731))
+* Initialize iOS project structure with Expo and React Native ([3223aee](https://github.com/Ucekay/react-native-nitro-text-input/commit/3223aeea4c992b26345a1b071886a74ad0e41b48))
+
+## 1.0.0 (2025-08-19)
+
+### Features
+
+* Add support for auto capitalization and font scaling in NitroTextInput ([15b97c8](https://github.com/Ucekay/react-native-nitro-text-input/commit/15b97c8675f334e6646a171a303e5f1ab98a243c))
+* **app:** update app structure and dependencies ([8721326](https://github.com/Ucekay/react-native-nitro-text-input/commit/8721326e79502ec48fffd6b6b88ef33a68d8f634))
+* Initialize Expo project with TypeScript ([f1c0344](https://github.com/Ucekay/react-native-nitro-text-input/commit/f1c034418cd87ac96471afb781219d80fd70b731))
+* Initialize iOS project structure with Expo and React Native ([3223aee](https://github.com/Ucekay/react-native-nitro-text-input/commit/3223aeea4c992b26345a1b071886a74ad0e41b48))
+
 ## 1.0.0-beta.1 (2025-08-19)
 
 ### Features